# docs/swagger.yaml
definitions:
  handlers.AssetResponse:
    properties:
      biz_line:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      snapshots:
        items:
          $ref: '#/definitions/handlers.SnapshotResponse'
        type: array
      state:
        type: string
      tags:
        items:
          type: string
        type: array
      updated_at:
        type: string
    type: object
...
  handlers.CreateAssetRequest:
    properties:
      biz_line:
        type: string
      content:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
...
  handlers.SnapshotResponse:
    properties:
      author:
        type: string
      commit_hash:
        type: string
      created_at:
        type: string
      eval_score:
        type: number
      reason:
        type: string
      version:
        type: string
    type: object
  handlers.UpdateAssetRequest:
    properties:
      biz_line:
        type: string
      description:
        type: string
      name:
        type: string
      state:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
paths:
  /api/v1/assets:
    get:
      consumes:
      - application/json
      description: Get all assets with optional filtering
      parameters:
      - description: Business line filter
        in: query
        name: biz_line
        type: string
      - description: Tag filter
        in: query
        name: tag
        type: string
      - description: State filter
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List assets
      tags:
      - assets
    post:
      consumes:
      - application/json
      description: Create a new asset
      parameters:
      - description: Asset creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateAssetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create asset
      tags:
      - assets
  /api/v1/assets/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an asset by ID
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete asset
      tags:
      - assets
    get:
      consumes:
      - application/json
      description: Get a single asset by its ID
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AssetResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get asset by ID
      tags:
      - assets
    put:
      consumes:
      - application/json
      description: Update an existing asset
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      - description: Asset update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateAssetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update asset
      tags:
      - assets
...
swagger: "2.0"